Git use custom component for groups
